+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Wednesday, June 14th 2023
+
+- Add `react-badges` ViteJS + React Web App  
+  (`make react`)
+
 ## Tuesday, June 13th 2023
 
 - Add support for rendering [PlantUML](https://plantuml.com/) charts on preview on GitPod